backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode[pil]==7.4.2
pillow==10.1.0
sendgrid==6.10.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0 